optimise: AT+CIUPDATE supports CDN network
In CDN network, the host request header of a HTTP GET should be filled with real hostname,
so that CDN servers can use this host item to request content from the origin server.
Otherwise, set an ip address of one CDN server as a host request header,
This will cause to origin server rejects the CDN request and returns 403 forbidden.